adding condition if the product isn't exists

+++ src/main/java/wsa/restfull1/ProductServiceController.java
@@ -58,16 +58,26 @@ public class ProductServiceController {
     
     @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
    public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) { 
-      productRepo.remove(id);
-      product.setId(id);
-      productRepo.put(id, product);
-      return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
+      if(!productRepo.containsKey(id)){
+          return new ResponseEntity<>("Product not found", HttpStatus.NOT_FOUND);
+      }
+      else {
+        productRepo.remove(id);
+        product.setId(id);
+        productRepo.put(id, product);
+        return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
+      }
    }
     
     @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<Object> delete(@PathVariable("id") String id) {
+        if(!productRepo.containsKey(id)){
+          return new ResponseEntity<>("Product not found", HttpStatus.NOT_FOUND);
+      }
+      else {
         productRepo.remove(id);
         return new ResponseEntity<>("Product is deleted successfully", HttpStatus.OK);
+      }
     }
     
 }
